requirements.txt
flask
requests
google-genai
pillow
python-dotenv
